# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Financial Dashboard API",
    "description": "REST API for health checks, ticker symbol search, and company overview data.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "getHealth",
        "summary": "Check API health",
        "tags": [
          "Health"
        ],
        "description": "Returns the current availability status for the API.",
        "responses": {
          "200": {
            "description": "API health status.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "API health status.",
                  "type": "object",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{symbol}/overview": {
      "get": {
        "operationId": "getCompanyOverview",
        "summary": "Get company overview",
        "tags": [
          "Companies"
        ],
        "description": "Returns company fundamentals and overview data for a ticker symbol.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "pattern": "^[A-Za-z0-9.-]+$"
            },
            "in": "path",
            "name": "symbol",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Company fundamentals and overview data.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Company fundamentals and overview data.",
                  "type": "object",
                  "required": [
                    "symbol",
                    "assetType",
                    "name",
                    "description",
                    "exchange",
                    "country",
                    "sector",
                    "industry",
                    "peRatio",
                    "eps",
                    "forwardPE"
                  ],
                  "properties": {
                    "symbol": {
                      "type": "string"
                    },
                    "assetType": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "exchange": {
                      "type": "string"
                    },
                    "country": {
                      "type": "string"
                    },
                    "sector": {
                      "type": "string"
                    },
                    "industry": {
                      "type": "string"
                    },
                    "peRatio": {
                      "type": "string"
                    },
                    "eps": {
                      "type": "string"
                    },
                    "forwardPE": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Requested resource was not found.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Unexpected server error.",
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/symbols/search": {
      "get": {
        "operationId": "searchSymbols",
        "summary": "Search ticker symbols",
        "tags": [
          "Symbols"
        ],
        "description": "Searches ticker symbols by keyword and returns matching securities.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "keywords",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Matching ticker symbol search results.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Matching ticker symbol search results.",
                  "type": "array",
                  "items": {
                    "description": "Ticker symbol search result.",
                    "type": "object",
                    "required": [
                      "symbol",
                      "name",
                      "type",
                      "region",
                      "currency"
                    ],
                    "properties": {
                      "symbol": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "region": {
                        "type": "string"
                      },
                      "currency": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Requested resource was not found.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Unexpected server error.",
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "API availability checks."
    },
    {
      "name": "Companies",
      "description": "Company fundamentals and overview data."
    },
    {
      "name": "Symbols",
      "description": "Ticker symbol search and validation."
    }
  ]
}